Revert "refactor(workspace): simplify document exports"
This reverts commit 644ce95106290d6b63df7d28e71dc9bb84d076aa.
