+++ apps/web/src/lib/markdown-to-docx.ts
@@ -38,22 +38,14 @@ import { parseMarkdownFrontmatter } from "@/components/workspace/markdown-frontm
 import { workspaceRemarkPlugins } from "@/components/workspace/markdown-plugins"
 import {
   getDocxDocumentAnchor,
+  getDocxDocumentTitle,
   getDocxHeadingAnchor,
+  getObsidianDocxLinkLabel,
+  resolveDocxInternalLink,
+  type DocxDocumentBundle,
+  type DocxSourceDocument,
 } from "@/lib/docx-document-bundle"
 import { findObsidianLinks } from "@/lib/kb-internal-links"
-import {
-  getMarkdownAstText,
-  getMarkdownDocumentTitle,
-  getObsidianMarkdownLinkLabel,
-  resolveMarkdownInternalLink,
-  type MarkdownDocumentBundle,
-  type MarkdownSourceDocument,
-} from "@/lib/markdown-document-bundle"
-import { getVegaLiteTitle, renderVegaLiteToSvg } from "@/lib/vega-lite-renderer"
-import {
-  getWorkspaceExportAbortReason,
-  withWorkspaceExportAbort,
-} from "@/lib/workspace-export-abort"
 
 type MarkdownNode = {
   type: string
@@ -89,7 +81,7 @@ type DocxRenderState = {
 }
 
 type DocxRenderContext = {
-  document: MarkdownSourceDocument
+  document: DocxSourceDocument
   headingOffset: number
   state: DocxRenderState
 }
@@ -114,7 +106,24 @@ function isSafeExternalUrl(value: string | undefined): value is string {
   }
 }
 
+function getVegaLiteTitle(spec: Record<string, unknown>): string {
+  const title = spec.title
+  if (typeof title === "string" && title.trim()) return title.trim()
+  if (!title || typeof title !== "object" || Array.isArray(title)) return "Vega-Lite chart"
+
+  const text = (title as Record<string, unknown>).text
+  if (typeof text === "string" && text.trim()) return text.trim()
+  if (Array.isArray(text)) {
+    const lines = text.filter((line): line is string => typeof line === "string")
+    if (lines.length > 0) return lines.join(" ")
+  }
+
+  return "Vega-Lite chart"
+}
+
 async function renderVegaLiteToPng(spec: Record<string, unknown>): Promise<RenderedChart> {
+  const vega = await import("vega")
+  const vegaLite = await import("vega-lite")
   const themeConfig = buildVegaConfig(FALLBACK_THEME)
   const specWithConfig = {
     ...spec,
@@ -129,41 +138,45 @@ async function renderVegaLiteToPng(spec: Record<string, unknown>): Promise<Rende
     },
     padding: CHART_PADDING,
   }
-  const svg = await renderVegaLiteToSvg(specWithConfig)
-  const rendered = await sharp(Buffer.from(svg), { density: 96 * CHART_RASTER_SCALE })
-    .trim({ background: "#ffffff", threshold: 10 })
-    .extend({
-      background: "#ffffff",
-      bottom: CHART_PADDING * CHART_RASTER_SCALE,
-      left: CHART_PADDING * CHART_RASTER_SCALE,
-      right: CHART_PADDING * CHART_RASTER_SCALE,
-      top: CHART_PADDING * CHART_RASTER_SCALE,
-    })
-    .resize({
-      fit: "inside",
-      height: (MAX_CHART_HEIGHT - CHART_PADDING * 2) * CHART_RASTER_SCALE,
-      width: (MAX_CHART_WIDTH - CHART_PADDING * 2) * CHART_RASTER_SCALE,
-      withoutEnlargement: true,
-    })
-    .png()
-    .toBuffer({ resolveWithObject: true })
+  const compiled = vegaLite.compile(specWithConfig as Parameters<typeof vegaLite.compile>[0])
+  const view = new vega.View(vega.parse(compiled.spec), { renderer: "none" })
 
-  return {
-    data: rendered.data,
-    height: Math.max(1, Math.round(rendered.info.height / CHART_RASTER_SCALE)),
-    title: getVegaLiteTitle(spec, "Vega-Lite chart"),
-    width: Math.max(1, Math.round(rendered.info.width / CHART_RASTER_SCALE)),
+  try {
+    const svg = await view.toSVG()
+    const rendered = await sharp(Buffer.from(svg), { density: 96 * CHART_RASTER_SCALE })
+      .trim({ background: "#ffffff", threshold: 10 })
+      .extend({
+        background: "#ffffff",
+        bottom: CHART_PADDING * CHART_RASTER_SCALE,
+        left: CHART_PADDING * CHART_RASTER_SCALE,
+        right: CHART_PADDING * CHART_RASTER_SCALE,
+        top: CHART_PADDING * CHART_RASTER_SCALE,
+      })
+      .resize({
+        fit: "inside",
+        height: (MAX_CHART_HEIGHT - CHART_PADDING * 2) * CHART_RASTER_SCALE,
+        width: (MAX_CHART_WIDTH - CHART_PADDING * 2) * CHART_RASTER_SCALE,
+        withoutEnlargement: true,
+      })
+      .png()
+      .toBuffer({ resolveWithObject: true })
+
+    return {
+      data: rendered.data,
+      height: Math.max(1, Math.round(rendered.info.height / CHART_RASTER_SCALE)),
+      title: getVegaLiteTitle(spec),
+      width: Math.max(1, Math.round(rendered.info.width / CHART_RASTER_SCALE)),
+    }
+  } finally {
+    view.finalize()
   }
 }
 
 async function renderVegaLiteCharts(
   nodes: MarkdownNode[],
   state: DocxRenderState,
-  signal?: AbortSignal,
 ): Promise<void> {
   for (const node of nodes) {
-    if (signal?.aborted) throw getWorkspaceExportAbortReason(signal)
-
     if (
       node.type === "code" &&
       node.lang === "vega-lite"
@@ -181,19 +194,15 @@ async function renderVegaLiteCharts(
       const spec = parseChartSpec(parsed)
       if (spec) {
         try {
-          node.chart = await withWorkspaceExportAbort(
-            renderVegaLiteToPng(spec),
-            signal,
-          )
+          node.chart = await renderVegaLiteToPng(spec)
           state.chartCount += 1
         } catch {
-          if (signal?.aborted) throw getWorkspaceExportAbortReason(signal)
           node.chart = undefined
         }
       }
     }
 
-    if (node.children) await renderVegaLiteCharts(node.children, state, signal)
+    if (node.children) await renderVegaLiteCharts(node.children, state)
   }
 }
 
@@ -202,7 +211,7 @@ function getIncludedLinkAnchor(
   syntax: "markdown" | "obsidian",
   context: DocxRenderContext,
 ): string | null {
-  const resolved = resolveMarkdownInternalLink(
+  const resolved = resolveDocxInternalLink(
     rawTarget,
     context.document.path,
     context.state.availablePaths,
@@ -233,7 +242,7 @@ function textChildren(
       children.push(new TextRun({ text: value.slice(cursor, link.from), ...style }))
     }
 
-    const label = getObsidianMarkdownLinkLabel(link.target)
+    const label = getObsidianDocxLinkLabel(link.target)
     const anchor = getIncludedLinkAnchor(link.target, "obsidian", context)
     children.push(
       anchor
@@ -432,8 +441,9 @@ function blockChildren(
       ]
       const sourceDepth = node.depth ?? 1
       const depth = Math.min(6, sourceDepth + (context?.headingOffset ?? 0))
-      const headingText = getMarkdownAstText(node)
-        .replace(/\s+/gu, " ")
+      const headingText = (node.children ?? [])
+        .map((child) => child.value ?? "")
+        .join("")
         .trim()
       const content = inlineChildren(node.children ?? [], {}, context)
       const headingChildren = context && headingText
@@ -542,7 +552,7 @@ function numberingLevels(format: (typeof LevelFormat)[keyof typeof LevelFormat])
   }))
 }
 
-function documentAnchorParagraph(document: MarkdownSourceDocument): Paragraph {
+function documentAnchorParagraph(document: DocxSourceDocument): Paragraph {
   return new Paragraph({
     children: [
       new Bookmark({
@@ -554,8 +564,8 @@ function documentAnchorParagraph(document: MarkdownSourceDocument): Paragraph {
   })
 }
 
-function getAppendixTitle(document: MarkdownSourceDocument, primaryTitle: string): string {
-  const documentTitle = getMarkdownDocumentTitle(document)
+function getAppendixTitle(document: DocxSourceDocument, primaryTitle: string): string {
+  const documentTitle = getDocxDocumentTitle(document)
   const primaryPrefix = `${primaryTitle} — `
   return documentTitle.startsWith(primaryPrefix)
     ? documentTitle.slice(primaryPrefix.length).trim()
@@ -563,47 +573,43 @@ function getAppendixTitle(document: MarkdownSourceDocument, primaryTitle: string
 }
 
 async function sourceDocumentChildren(
-  document: MarkdownSourceDocument,
+  document: DocxSourceDocument,
   headingOffset: number,
   state: DocxRenderState,
-  signal?: AbortSignal,
 ): Promise<FileChild[]> {
-  if (signal?.aborted) throw getWorkspaceExportAbortReason(signal)
-
   const frontmatter = parseMarkdownFrontmatter(document.markdown)
   let processor = unified().use(remarkParse)
   for (const plugin of workspaceRemarkPlugins) processor = processor.use(plugin)
   const tree = processor.parse(frontmatter.body) as MarkdownNode
   const children = tree.children ?? []
-  const documentTitle = getMarkdownDocumentTitle(document)
+  const documentTitle = getDocxDocumentTitle(document)
 
   if (headingOffset > 0) {
     const titleHeadingIndex = children.findIndex(
       (node) =>
         node.type === "heading" &&
         node.depth === 1 &&
-        getMarkdownAstText(node).replace(/\s+/gu, " ").trim() === documentTitle,
+        (node.children ?? []).map((child) => child.value ?? "").join("").trim() === documentTitle,
     )
     if (titleHeadingIndex >= 0) children.splice(titleHeadingIndex, 1)
   }
 
-  await renderVegaLiteCharts(children, state, signal)
+  await renderVegaLiteCharts(children, state)
   const context: DocxRenderContext = { document, headingOffset, state }
   return children.flatMap((node) => blockChildren(node, 0, context))
 }
 
 export async function markdownToDocx(
-  input: string | MarkdownDocumentBundle,
-  options?: { signal?: AbortSignal },
+  input: string | DocxDocumentBundle,
 ): Promise<Buffer> {
-  const bundle: MarkdownDocumentBundle = typeof input === "string"
+  const bundle: DocxDocumentBundle = typeof input === "string"
     ? {
         appendices: [],
         availablePaths: ["article.md"],
         primary: { markdown: input, path: "article.md" },
       }
     : input
-  const title = getMarkdownDocumentTitle(bundle.primary)
+  const title = getDocxDocumentTitle(bundle.primary)
   const state: DocxRenderState = {
     availablePaths: bundle.availablePaths,
     chartCount: 0,
@@ -611,12 +617,7 @@ export async function markdownToDocx(
       [bundle.primary, ...bundle.appendices].map((document) => document.path.toLowerCase()),
     ),
   }
-  const primaryChildren = await sourceDocumentChildren(
-    bundle.primary,
-    0,
-    state,
-    options?.signal,
-  )
+  const primaryChildren = await sourceDocumentChildren(bundle.primary, 0, state)
   const sections: ISectionOptions[] = [
     {
       properties: {
@@ -629,9 +630,9 @@ export async function markdownToDocx(
   ]
 
   for (const appendix of bundle.appendices) {
-    const sourceTitle = getMarkdownDocumentTitle(appendix)
+    const sourceTitle = getDocxDocumentTitle(appendix)
     const appendixTitle = `Appendix. ${getAppendixTitle(appendix, title)}`
-    const content = await sourceDocumentChildren(appendix, 1, state, options?.signal)
+    const content = await sourceDocumentChildren(appendix, 1, state)
     sections.push({
       properties: {
         type: SectionType.NEXT_PAGE,
