+++ apps/web/src/lib/markdown-to-pdf-html.ts
@@ -21,21 +21,14 @@ import {
 } from "@/components/workspace/markdown-plugins"
 import { findObsidianLinks } from "@/lib/kb-internal-links"
 import {
+  getObsidianPdfLinkLabel,
   getPdfDocumentAnchor,
+  getPdfDocumentTitle,
   getPdfHeadingAnchor,
+  resolvePdfInternalLink,
+  type PdfDocumentBundle,
+  type PdfSourceDocument,
 } from "@/lib/pdf-document-bundle"
-import {
-  getMarkdownDocumentTitle,
-  getObsidianMarkdownLinkLabel,
-  resolveMarkdownInternalLink,
-  type MarkdownDocumentBundle,
-  type MarkdownSourceDocument,
-} from "@/lib/markdown-document-bundle"
-import {
-  getVegaLiteTitle,
-  renderVegaLiteToSvg as renderVegaLiteSpecToSvg,
-} from "@/lib/vega-lite-renderer"
-import { withWorkspaceExportAbort } from "@/lib/workspace-export-abort"
 
 const MAX_VEGA_CHARTS = 20
 const MAX_DATA_IMAGE_BYTES = 4 * 1024 * 1024
@@ -64,15 +57,50 @@ type PdfRenderState = {
 }
 
 type PdfDocumentRenderContext = {
-  document: MarkdownSourceDocument
+  document: PdfSourceDocument
   headingOffset: number
   state: PdfRenderState
 }
 
+function getAbortReason(signal: AbortSignal): Error {
+  if (signal.reason instanceof Error) return signal.reason
+  return new Error("pdf_export_aborted")
+}
+
+function withAbort<T>(operation: Promise<T>, signal?: AbortSignal): Promise<T> {
+  if (!signal) return operation
+  if (signal.aborted) return Promise.reject(getAbortReason(signal))
+
+  return new Promise<T>((resolve, reject) => {
+    const abort = () => reject(getAbortReason(signal))
+    signal.addEventListener("abort", abort, { once: true })
+    operation.then(
+      (value) => {
+        signal.removeEventListener("abort", abort)
+        resolve(value)
+      },
+      (error: unknown) => {
+        signal.removeEventListener("abort", abort)
+        reject(error)
+      },
+    )
+  })
+}
+
 async function renderVegaLiteToSvg(spec: Record<string, unknown>): Promise<string> {
+  const vega = await import("vega")
+  const vegaLite = await import("vega-lite")
+
   const config = buildVegaConfig(FALLBACK_THEME)
   const specWithConfig = { ...spec, config }
-  return renderVegaLiteSpecToSvg(specWithConfig)
+  const compiled = vegaLite.compile(specWithConfig as Parameters<typeof vegaLite.compile>[0])
+  const runtime = vega.parse(compiled.spec)
+  const view = new vega.View(runtime, { renderer: "none" })
+  try {
+    return await view.toSVG()
+  } finally {
+    view.finalize()
+  }
 }
 
 type VegaLiteTarget = {
@@ -103,6 +131,21 @@ function extractVegaLiteSpec(preNode: Element): Record<string, unknown> | null {
   }
 
   return parseChartSpec(parsed)
+}
+
+function getVegaLiteTitle(spec: Record<string, unknown>): string {
+  const title = spec.title
+  if (typeof title === "string" && title.trim()) return title.trim()
+  if (!title || typeof title !== "object" || Array.isArray(title)) return "Untitled figure"
+
+  const text = (title as Record<string, unknown>).text
+  if (typeof text === "string" && text.trim()) return text.trim()
+  if (Array.isArray(text)) {
+    const lines = text.filter((line): line is string => typeof line === "string")
+    if (lines.length > 0) return lines.join(" ")
+  }
+
+  return "Untitled figure"
 }
 
 function normalizeFigureLabel(label: string): string {
@@ -185,7 +228,7 @@ function rehypeVegaLiteToSvg(state: PdfRenderState, signal?: AbortSignal) {
       state.vegaChartCount += 1
       let svg: string | null = null
       try {
-        svg = await withWorkspaceExportAbort(renderVegaLiteToSvg(target.spec), signal)
+        svg = await withAbort(renderVegaLiteToSvg(target.spec), signal)
       } catch {
         continue
       }
@@ -200,7 +243,7 @@ function rehypeVegaLiteToSvg(state: PdfRenderState, signal?: AbortSignal) {
               children: [{ type: "raw", value: svg } as unknown as ElementContent],
             },
           ],
-          getVegaLiteTitle(target.spec, "Untitled figure"),
+          getVegaLiteTitle(target.spec),
         ),
       }
     }
@@ -223,7 +266,7 @@ function removeAppendixDocumentTitleHeading(
 ): void {
   if (context.headingOffset === 0) return
 
-  const documentTitle = getMarkdownDocumentTitle(context.document)
+  const documentTitle = getPdfDocumentTitle(context.document)
     .replace(/\s+/gu, " ")
     .trim()
   const initialContentIndex = tree.children.findIndex(
@@ -244,7 +287,7 @@ function getIncludedLinkHref(
   syntax: "markdown" | "obsidian",
   context: PdfDocumentRenderContext,
 ): string | null {
-  const resolved = resolveMarkdownInternalLink(
+  const resolved = resolvePdfInternalLink(
     rawTarget,
     context.document.path,
     context.state.availablePaths,
@@ -270,7 +313,7 @@ function rewriteMarkdownLinks(
       const href = child.properties.href
       if (typeof href !== "string") continue
 
-      const resolved = resolveMarkdownInternalLink(
+      const resolved = resolvePdfInternalLink(
         href,
         context.document.path,
         context.state.availablePaths,
@@ -325,7 +368,7 @@ function rewriteObsidianLinks(
         })
       }
 
-      const label = getObsidianMarkdownLinkLabel(link.target)
+      const label = getObsidianPdfLinkLabel(link.target)
       const href = getIncludedLinkHref(link.target, "obsidian", context)
       if (href) {
         replacements.push({
@@ -767,10 +810,10 @@ function escapeHtml(value: string): string {
 }
 
 function getAppendixSectionTitle(
-  document: MarkdownSourceDocument,
+  document: PdfSourceDocument,
   primaryTitle: string,
 ): string {
-  const documentTitle = getMarkdownDocumentTitle(document)
+  const documentTitle = getPdfDocumentTitle(document)
   const primaryTitlePrefix = `${primaryTitle} — `
 
   return documentTitle.startsWith(primaryTitlePrefix)
@@ -779,7 +822,7 @@ function getAppendixSectionTitle(
 }
 
 async function renderPdfSourceDocument(
-  document: MarkdownSourceDocument,
+  document: PdfSourceDocument,
   headingOffset: number,
   state: PdfRenderState,
   signal?: AbortSignal,
@@ -799,7 +842,7 @@ async function renderPdfSourceDocument(
     .use(() => rehypePdfDocumentSemantics(context))
     .use(rehypeStringify, { allowDangerousHtml: true })
 
-  const result = await withWorkspaceExportAbort(processor.process(frontmatter.body), signal)
+  const result = await withAbort(processor.process(frontmatter.body), signal)
   return String(result)
 }
 
@@ -844,10 +887,10 @@ function buildTableOfFigures(figures: PdfFigureEntry[]): string {
 }
 
 export async function markdownToPdfHtml(
-  bundle: MarkdownDocumentBundle,
+  bundle: PdfDocumentBundle,
   options?: { logoBase64?: string; signal?: AbortSignal },
 ): Promise<string> {
-  const title = getMarkdownDocumentTitle(bundle.primary)
+  const title = getPdfDocumentTitle(bundle.primary)
   const state: PdfRenderState = {
     availablePaths: bundle.availablePaths,
     figureCount: 0,
@@ -879,7 +922,7 @@ export async function markdownToPdfHtml(
       state,
     )
     const documentTitleHeadingId = getUniquePdfId(
-      getPdfHeadingAnchor(appendix.path, getMarkdownDocumentTitle(appendix)),
+      getPdfHeadingAnchor(appendix.path, getPdfDocumentTitle(appendix)),
       state,
     )
     state.headings.push({
